[IS-1290] separate block_sync from block_manager
